requirements.txt
absl-py==1.4.0
alembic==1.9.2
antlr4-python3-runtime==4.9.3
attrs==22.2.0
autopage==0.5.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
cliff==4.1.0
cloudpickle==2.2.0
cmaes==0.9.1
cmd2==2.4.2
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
grpcio==1.51.1
gym==0.21.0
hydra-core==1.3.1
hydra-optuna-sweeper==1.2.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.2
jax-jumpy==0.2.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
optuna==2.10.1
packaging==23.0
pandas==1.5.2
pbr==5.11.1
Pillow==9.4.0
prettytable==3.6.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
Shimmy==0.2.0
six==1.16.0
SQLAlchemy==1.4.46
stable-baselines3==1.7.0
stevedore==4.1.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.11.0
